Downloads/MetaBitDepositBot-main/MetaBitDepositBot-mainV1/MetaBitDepositBot-main/web_dashboard31-12/requirements.txt
python-binance==1.0.19
python-telegram-bot==20.7
python-dotenv==1.0.0
backoff==2.2.1
urllib3==2.0.7
requests==2.31.0
asyncio==3.4.3
web3==6.8.0
aiosqlite
python-telegram-bot
aiomysql
asyncpg
backoff
datadog
autopep8
psutil
flask==3.0.0
flask-login==0.6.3
aiohttp==3.9.1
pandas==2.1.4
plotly==5.18.0
plotly-express==0.4.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
gunicorn==21.2.0
eth_utils
